Use angular from bower instead of downloading.
add bower.json and reference to angular.

        bower init
        bower install --save angular

add the grunt task grunt-bowercopy
and added to the list of tasks that lineman knows about.

created a bower section in config/application.js that copies the bowere
angular files into the vendor directory.

get rid of the cdn reference.  Since angular is now loaded at the end
of the file instead of the beginning, the app would flicker as angular
bootstraps itself.  So add the ng-cloak directive and manually add the
style to style.css