Add in angular-route
lots of problems.. used bower to install angular-route and modified
config/files.js to force angular to get sourced in first.

getting the route to work forces you into some decisions... to html5 or
not to html5.

if you use html5, then you need to have some rewrite rules on your web
server to handle the case someone hits the refresh button.
Sometime back the angular team added "#!" to the non html5 url instead
of just "#", and it broke some things.

I think in the future, links should not just link to urls, but should
execute a "$location.goto('/someurl') in the routeProvider.  time will
tell.

Anyway, a few examples of routes, including one route that is slow to
respond.

with the current method of using links, you don't get feedback that the
system is working... maybe by listening to events you could throw up a
"working.." spinner. but that's for tomorrow.