move booklist to services directory.
tiny bit of bad use of inline css to make it look a tiny bit better.
(min-height for the list of books and 2 columns for book/book image).

also added the standard ng-repeat/serach box, but turns out that there
is an issue with ui-sref-active and ng-repeat.  There is a race
condition and the active class isn't added at the right time. Had to fix
with a horrible $timeout.. <sigh>.