Add in some simple security stuff.
Add a list of  users in server.js, along with login/logout.
only cookie based right now.
some code to handle expiration times.

Lots of ugly code in the main controller to handle when someone
lands on our page without a valid cookie, and is inside of our app
trying to view a protected resource.

if you try and link to http://localhost:8000/#!/books/4

the server needs to request the booklist, and book 4.

If you are not logged in, you get an access denied.

The way I handle the login page, the system will successfully navigate
to the books/book state, and fire off those requests that will fail,
while showing the login page.

when you log in, the page is displayed with the "you don't have
permission" error messages.

To keep this from happening, store the state on the root scope, detect
when changing to non index state without a logged in user and redirecrt
to index, while storing where they wanted to go.
after login (or whoami call), navigate to that saved value.

Will change this next time to just make a root state that has the user
as part of the resolve block.  That can fix the "hitting refresh while
you are logged in" problem, but maybe not the navigating while you
aren't logged in yet... need to think about it more.